# backend/app/api/static/docs/openapi-3.yaml
openapi: 3.0.0
info:
  description: Track, Manage, and Organize your Things.
  title: HomeBot API
  contact:
    name: HomeBot Team
    url: https://github.com/pritish-codes/homebot
  version: "1.0"
paths:
  /v1/actions/create-missing-thumbnails:
    post:
      security:
        - Bearer: []
      description: Creates thumbnails for items that are missing them
      tags:
        - Actions
      summary: Create Missing Thumbnails
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
  /v1/actions/ensure-asset-ids:
    post:
      security:
        - Bearer: []
      description: Ensures all items in the database have an asset ID
      tags:
        - Actions
      summary: Ensure Asset IDs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
  /v1/actions/ensure-import-refs:
    post:
      security:
        - Bearer: []
      description: Ensures all items in the database have an import ref
      tags:
        - Actions
      summary: Ensures Import Refs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
  /v1/actions/set-primary-photos:
    post:
      security:
        - Bearer: []
      description: Sets the first photo of each item as the primary photo
      tags:
        - Actions
      summary: Set Primary Photos
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
  /v1/actions/wipe-inventory:
    post:
      security:
        - Bearer: []
      description: Deletes all items in the inventory
      tags:
        - Actions
      summary: Wipe Inventory
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v1.WipeInventoryOptions"
        description: Wipe options
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
  /v1/actions/zero-item-time-fields:
    post:
      security:
        - Bearer: []
      description: Resets all item date fields to the beginning of the day
      tags:
        - Actions
      summary: Zero Out Time Fields
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/v1.ActionAmountResult"
  "/v1/assets/{id}":
    get:
      security:
        - Bearer: []
      tags:
        - Items
      summary: Get Item by Asset ID
      parameters:
        - description: Asset ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.PaginationResult-repo_EntitySummary"
...
servers:
  - url: /api
components:
  securitySchemes:
    Bearer:
      description: "\"Type 'Bearer TOKEN' to correctly set the API Key\""
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    repo.EntitySummary:
      type: object
      properties:
        archived:
          type: boolean
        assetId:
          type: string
          example: "0"
        createdAt:
          type: string
        description:
          type: string
        entityType:
          allOf:
            - $ref: "#/components/schemas/repo.EntityTypeSummary"
          x-omitempty: true
          nullable: true
        id:
          type: string
        imageId:
          type: string
          x-omitempty: true
          nullable: true
        insured:
          type: boolean
        itemCount:
          description: Container-specific (populated when querying locations)
          type: number
        name:
          type: string
        parent:
          description: Edges
          allOf:
            - $ref: "#/components/schemas/repo.EntitySummary"
          x-omitempty: true
          nullable: true
        purchasePrice:
          type: number
        quantity:
          type: number
        soldDate:
          description: Sale details
          type: string
        tags:
          type: array
          items:
            $ref: "#/components/schemas/repo.TagSummary"
        thumbnailId:
          type: string
          x-omitempty: true
          nullable: true
        updatedAt:
          type: string
...
    repo.EntityTemplateSummary:
      type: object
      properties:
        createdAt:
          type: string
        description:
          type: string
        id:
          type: string
        name:
          type: string
        updatedAt:
          type: string
...
    repo.EntityTypeSummary:
      type: object
      properties:
        createdAt:
          type: string
        defaultTemplate:
          $ref: "#/components/schemas/repo.EntityTemplateSummary"
        defaultTemplateId:
          type: string
        description:
          type: string
        icon:
          type: string
        id:
          type: string
        isLocation:
          type: boolean
        name:
          type: string
        updatedAt:
          type: string
...
    repo.PaginationResult-repo_EntitySummary:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/repo.EntitySummary"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...
    repo.TagSummary:
      type: object
      properties:
        color:
          type: string
        createdAt:
          type: string
        description:
          type: string
        icon:
          type: string
        id:
          type: string
        name:
          type: string
        parentId:
          type: string
          nullable: true
        updatedAt:
          type: string
...
    v1.ActionAmountResult:
      type: object
      properties:
        completed:
          type: integer
...
    v1.WipeInventoryOptions:
      type: object
      properties:
        wipeLocations:
          type: boolean
        wipeMaintenance:
          type: boolean
        wipeTags:
          type: boolean
...